+++ sw/applications/src/kernels/mnist_fp64_network.h
@@ -8,11 +8,11 @@
 
 void feedforward_fp64n(uint32_t IN_CH1, uint32_t IN_CH2, uint32_t OUT_CH, 
                 double *weights, uint32_t ldW, double *biases, double* activations,
-                uint32_t ldB, float *image, uint32_t ldI, uint32_t compute_id);
+                uint32_t ldB, double *image, uint32_t ldI, uint32_t compute_id);
 
 void softmax_activation_fp64n(uint32_t IN_CH1, uint32_t IN_CH2, uint32_t OUT_CH, 
                 double *weights, uint32_t ldW, double *activations, uint32_t ldB,
-                float *image, uint32_t ldI, uint32_t compute_id, 
+                double *image, uint32_t ldI, uint32_t compute_id, 
                 uint32_t compute_num, double *max);
 
 void gradient_update_fp64(uint32_t IN_CH1, uint32_t IN_CH2, uint32_t OUT_CH,
